# specs/001-ghostarr-newsletter/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ghostarr API
  description: Newsletter generator for media servers
  version: 1.0.0
  contact:
    name: Ghostarr
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /templates:
    get:
      tags: [templates]
      summary: List templates
      operationId: listTemplates
      responses:
        '200':
          description: Template list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateResponse'

    post:
      tags: [templates]
      summary: Create template
      operationId: createTemplate
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, file]
              properties:
                name:
                  type: string
                description:
                  type: string
                tags:
                  type: string
                  description: JSON array of tags
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
  /templates/{id}:
    get:
      tags: [templates]
      summary: Get template
      operationId: getTemplate
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [templates]
      summary: Update template
      operationId: updateTemplate
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'

    delete:
      tags: [templates]
      summary: Delete template
      operationId: deleteTemplate
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
        '409':
          description: Template in use by active schedules
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                  schedules:
                    type: array
                    items:
                      type: string
  /templates/{id}/preview:
    get:
      tags: [templates]
      summary: Preview template with mock data
      operationId: previewTemplate
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Preview HTML
          content:
            application/json:
              schema:
                type: object
                properties:
                  html:
                    type: string
...
  /settings:
    get:
      tags: [settings]
      summary: Get all settings
      operationId: getSettings
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'

    put:
      tags: [settings]
      summary: Update settings
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
  /settings/services/{service}/test:
    post:
      tags: [settings]
      summary: Test service connection
      operationId: testService
      parameters:
        - name: service
          in: path
          required: true
          schema:
            type: string
            enum: [tautulli, tmdb, ghost, romm, komga, audiobookshelf, tunarr]
      responses:
        '200':
          description: Connection result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  version:
                    type: string
...
  /preferences:
    get:
      tags: [settings]
      summary: Get user preferences
      operationId: getPreferences
      responses:
        '200':
          description: Preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferenceResponse'

    put:
      tags: [settings]
      summary: Update user preferences
      operationId: updatePreferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferenceUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferenceResponse'
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
...
  schemas:
...
    GenerationConfig:
      type: object
      required: [template_id]
      properties:
        template_id:
          type: string
          format: uuid
        title:
          type: string
          default: "Newsletter {{date.week}}"
        publication_mode:
          type: string
          enum: [draft, publish, email, email+publish]
          default: draft
        ghost_newsletter_id:
          type: string
          nullable: true
        tautulli:
          $ref: '#/components/schemas/TautulliConfig'
        romm:
          $ref: '#/components/schemas/ContentSourceConfig'
        komga:
          $ref: '#/components/schemas/ContentSourceConfig'
        audiobookshelf:
          $ref: '#/components/schemas/ContentSourceConfig'
        tunarr:
          $ref: '#/components/schemas/TunarrConfig'
        statistics:
          $ref: '#/components/schemas/StatisticsConfig'
        maintenance:
          $ref: '#/components/schemas/MaintenanceConfig'
        max_total_items:
          type: integer
          default: -1
        skip_if_empty:
          type: boolean
          default: false
    ContentSourceConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        days:
          type: integer
          minimum: 1
          maximum: 90
          default: 7
        max_items:
          type: integer
          minimum: -1
          default: -1
    TautulliConfig:
      allOf:
        - $ref: '#/components/schemas/ContentSourceConfig'
        - type: object
          properties:
            featured_item:
              type: boolean
              default: false
    TunarrConfig:
      allOf:
        - $ref: '#/components/schemas/ContentSourceConfig'
        - type: object
          properties:
            days:
              type: integer
              minimum: 1
              maximum: 7
              default: 7
            channels:
              type: array
              items:
                type: string
            display_format:
              type: string
              enum: [grid, list]
              default: grid
    StatisticsConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        days:
          type: integer
          minimum: 1
          maximum: 90
          default: 7
        include_comparison:
          type: boolean
          default: false
    MaintenanceConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        description:
          type: string
        type:
          type: string
          enum: [scheduled, outage, network, update, improvement, security]
          default: scheduled
        duration_value:
          type: integer
          default: 1
        duration_unit:
          type: string
          enum: [hours, days, weeks]
          default: hours
        start_datetime:
          type: string
          format: date-time
          nullable: true

    # ---- Templates ----
    TemplateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        file_path:
          type: string
        preset_config:
          $ref: '#/components/schemas/GenerationConfig'
        is_default:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TemplateUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        preset_config:
          $ref: '#/components/schemas/GenerationConfig'

    # ---- Schedules ----
...
    SettingsResponse:
      type: object
      properties:
        services:
          type: object
          properties:
            tautulli:
              $ref: '#/components/schemas/ServiceConfig'
            tmdb:
              $ref: '#/components/schemas/ServiceConfig'
            ghost:
              $ref: '#/components/schemas/ServiceConfig'
            romm:
              $ref: '#/components/schemas/ServiceConfig'
            komga:
              $ref: '#/components/schemas/ServiceConfig'
            audiobookshelf:
              $ref: '#/components/schemas/ServiceConfig'
            tunarr:
              $ref: '#/components/schemas/ServiceConfig'
        retention:
          type: object
          properties:
            history_days:
              type: integer
            logs_days:
              type: integer
        notifications:
          type: object
          properties:
            admin_email:
              type: string
              nullable: true
    SettingsUpdate:
      type: object
      properties:
        services:
          type: object
          additionalProperties:
            type: object
            properties:
              url:
                type: string
              api_key:
                type: string
        retention:
          type: object
          properties:
            history_days:
              type: integer
            logs_days:
              type: integer
        notifications:
          type: object
          properties:
            admin_email:
              type: string
    ServiceConfig:
      type: object
      properties:
        url:
          type: string
        configured:
          type: boolean
        status:
          type: string
          enum: [connected, error, unconfigured]
        last_checked:
          type: string
          format: date-time
          nullable: true

    # ---- User Preferences ----
    UserPreferenceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        theme:
          type: string
          enum: [light, dark, system]
        language:
          type: string
        timezone:
          type: string
    UserPreferenceUpdate:
      type: object
      properties:
        theme:
          type: string
          enum: [light, dark, system]
        language:
          type: string
        timezone:
          type: string

    # ---- Logs ----
...